Refactored Using Context
Ready for merge
