+++ src/Header/Header.css
@@ -1,8 +1,22 @@
 header {
-    background-color: white;
+    background-color: rgba(255, 255, 255, 0);
     width: 100%;
 }
 
 .logo {
-    width: 50vw;
+    width: 50%;
+}
+
+@media only screen and (min-width: 990px) {
+    .logo {
+        width: 15%;
+        align-self: left;
+    }
+
+    header {
+        display: flex;
+        justify-content: space-between;
+        align-items: flex-end;
+        padding: 0 5vw;
+    }
 }
