+++ src/Header/Header.js
@@ -1,13 +1,15 @@
-import React, { Component } from 'react';
-import logo from '../images/weather-watch-logo.png';
-import './Header.css';
+import React, { Component } from "react";
+import {CurrentContext} from "../CurrentContext";
+import logo from "../images/weather-watch-logo.png";
+import "./Header.css";
 
 export default class Header extends Component {
-    render() {
-        return (
-            <header>
-                <img src={logo} alt="Logo" className='logo' />
-            </header>
-        )
-    }
+  static contextType = CurrentContext;
+  render() {
+    return (
+      <header>
+        <img src={logo} alt="Logo" className="logo" />
+      </header>
+    );
+  }
 }
